Add README.md for APK installation, update .idea project files
Add a README with instructions for installing APK for different Android versions to clarify the project setup process for new users. Also updated .idea project files (misc and vcs.xml). The misc.xml update removes the xml version line, as it's unnecessary and was causing parsing issues. The vcs.xml was added for better git compatibility in the IntelliJ project setup.